clockify-auth-server/requirements.txt
boto3==1.16.34
botocore==1.19.34
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.5
requests==2.25.0
s3transfer==0.3.3
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
